backend/requirements.txt
black==24.8.0
Django==3.2.3
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.0.0